Fix cart hours update bug.
Previously, the number of hours per week in the cart was not updating when switching between calendars due to the fact that the hours prop pulled its information from the fullcalendar object's list of events. Unlike the redux state of the app, the fullcalendar object's list of events were not attached to the update lifecycle of the components, and therefore did not update correctly. The CartContainer now receives its number of hours information from the state, as is appropriate.