Move remove calendar button to calendar dropdown
Previously, the remove calendar button was located at the top right hand corner of the calendar. Through testing it was determined that many users could not easily find the button, as it was far away from the add calendar button. The remove calendar button is now moved inside the calendar chip itself. The remove button only appears if there is more than 1 calendar present.