Use section topic as title when appropriate
Most ordinary sections contain a field called 'topic' that has a value of empty string. Some sections, however, for courses such as Special Topics in EECS or Special Topics in Theatre, they use topics to distinguish between sections and these topic names colloquially serve as the course titles. Thus, whenever the topic field is not empty string, the sections selection list displays it and the calendar event displays it as the title.