Make "Serif.nu" on top left clickable
The "Serif.nu" on the top left hand corner of the screen (just to the right of the hamburger menu button on the App Bar) was previously not clickable. Now, it acts as a link to the home page of the app.

The Serif.nu menu item has been removed from the Nav menu.