Put home page link in nav bar
There is now a link to return to the home page in the Navbar (the logo still serves as a link to the home page as well). This gives the user the choice between either option - whichever one makes more sense to them to use in order to return to the home page.